# server/openapi.yaml
openapi: 3.0.3
info:
  title: Awesome Tracker API
  version: 1.0.0
  description: >
    Backend for the Awesome Tracker app: JWT auth, profile/avatar
    management, and chat. See server/README.md for local setup.
  license:
    name: MIT
    url: https://github.com/KhurshidbekUrazaliev/awesome-tracker-app/blob/main/LICENSE
servers:
  - url: http://localhost:3000/api
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        avatar: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, createdAt]

    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        content: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, conversationId, senderId, content, createdAt]

    Conversation:
      type: object
      properties:
        id: { type: string, format: uuid }
        participants:
          type: array
          items: { type: string, format: uuid }
        lastMessage:
          allOf: [{ $ref: '#/components/schemas/Message' }]
          description: Omitted if the conversation has no messages yet.
        unreadCount: { type: integer }
        updatedAt: { type: string, format: date-time }
      required: [id, participants, unreadCount, updatedAt]

    Error:
      type: object
      properties:
        message: { type: string }
      required: [message]

  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but not permitted to perform this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Liveness probe
      security: []
      responses:
        '200': { description: OK }

  /ready:
    get:
      summary: Readiness probe (checks the database is reachable)
      security: []
      responses:
        '200': { description: Ready }
        '503': { description: Not ready }

  /auth/signup:
    post:
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
              required: [name, email, password]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { description: Email already registered }

  /auth/login:
    post:
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Invalid email or password }

  /auth/logout:
    post:
      summary: Log out (stateless JWT — no-op server-side)
      security: []
      responses:
        '204': { description: Logged out }

  /auth/refresh:
    post:
      summary: Reissue a token for the current user
      responses:
        '200':
          description: New token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/forgot-password:
    post:
      summary: Request a password reset (logs a reset token server-side; no email provider configured)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: "Always 200, to avoid account enumeration" }

  /auth/reset-password:
    post:
      summary: Reset password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
              required: [token, password]
      responses:
        '200': { description: Password updated }
        '400': { description: Invalid or expired reset link }

  /users/me:
    get:
      summary: Get the current user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      summary: Update the current user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                avatar: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      summary: Delete the current user's account
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /users/{id}:
    get:
      summary: Get a user by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  /users/me/avatar:
    post:
      summary: Upload an avatar image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar: { type: string, format: binary }
      responses:
        '200':
          description: Avatar updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  avatar: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /upload:
    post:
      summary: Upload an arbitrary file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  fileName: { type: string }
                  fileSize: { type: integer }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /chat/conversations:
    get:
      summary: List the current user's conversations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Conversation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participantIds:
                  type: array
                  items: { type: string, format: uuid }
              required: [participantIds]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /chat/conversations/{id}/messages:
    get:
      summary: List messages in a conversation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Send a message
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string, minLength: 1 }
              required: [content]
      responses:
        '201':
          description: Sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /chat/conversations/{id}/read:
    post:
      summary: Mark a conversation as read for the current user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Marked read }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /chat/messages/{id}:
    delete:
      summary: Delete a message (sender only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }